# ShopAPI/ShopAPI/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My Shop API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Products": {
      "get": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "category",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "minPrice",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "maxPrice",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "sort",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "order",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Name",
            "required": true,
            "type": "string",
            "maxLength": 100
          },
          {
            "in": "formData",
            "name": "Brand",
            "required": true,
            "type": "string",
            "maxLength": 100
          },
          {
            "in": "formData",
            "name": "Category",
            "required": true,
            "type": "string",
            "maxLength": 100
          },
          {
            "in": "formData",
            "name": "Price",
            "required": true,
            "type": "number",
            "format": "double"
          },
          {
            "in": "formData",
            "name": "Description",
            "type": "string",
            "maxLength": 4000
          },
          {
            "in": "formData",
            "name": "ImageFile",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Name",
            "required": true,
            "type": "string",
            "maxLength": 100
          },
          {
            "in": "formData",
            "name": "Brand",
            "required": true,
            "type": "string",
            "maxLength": 100
          },
          {
            "in": "formData",
            "name": "Category",
            "required": true,
            "type": "string",
            "maxLength": 100
          },
          {
            "in": "formData",
            "name": "Price",
            "required": true,
            "type": "number",
            "format": "double"
          },
          {
            "in": "formData",
            "name": "Description",
            "type": "string",
            "maxLength": 4000
          },
          {
            "in": "formData",
            "name": "ImageFile",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Users": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    }
  },
...
  "securityDefinitions": {
    "oauth2": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Please enter token"
    }
  }
}